103-fibonacci.c

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    unsigned long long first = 1;
-    unsigned long long second = 2;
-    unsigned long long next;
-    unsigned long long sum = 2; // Start with the sum as 2 (since the second term is even)
+    unsigned long first = 1;
+    unsigned long second = 2;
+    unsigned long next;
+    unsigned long sum = 2; // Start with the sum as 2 (since the second term is even)
 
     printf("1, 2, "); // Print the first two terms of the Fibonacci sequence
 
@@ -16,12 +16,12 @@ int main(void)
         {
             sum += next;
         }
-        printf("%llu, ", next); // Print the next term
+        printf("%lu, ", next); // Print the next term
         first = second;
         second = next;
     }
 
-    printf("\nSum of even-valued terms: %llu\n", sum); // Print the sum of even-valued terms
+    printf("\nSum of even-valued terms: %lu\n", sum); // Print the sum of even-valued terms
 
     return 0;
 }
